Feat: 방명록 검색 기능 추가

+++ src/main/java/com/babzip/backend/guestbook/dto/response/GuestbookResponseDto.java
@@ -6,7 +6,7 @@ public record GuestbookResponseDto(
         String restaurantName,
         String kakaoPlaceId,
         String content,
-        Integer rating
+        Long rating
 ) {
     public static GuestbookResponseDto toDto(Guestbook guestbook) {
         return new GuestbookResponseDto(
